Add /teams/:teamId/sites/:siteId/stories client-side

+++ packages/client/src/crud/CRUDList.tsx
@@ -5,25 +5,42 @@ import {Link} from '../components'
 
 export interface ICRUDListProps<T> {
   nameKey: keyof T
-  editLink: (item: T) => string
+  editLink?: (item: T) => string
   itemIds: ReadonlyArray<number>
   itemsById: Record<number, T>
-  newLink: string
-  onRemove: (t: T) => void
+  newLink?: string
+  onRemove?: (t: T) => void
   title: string
+  Info?: React.ComponentType<ICRUDItemInfoProps<T>>
 }
 
 export interface ICRUDItemRowProps<T> {
+  Info?: React.ComponentType<ICRUDItemInfoProps<T>>
   nameKey: keyof T
-  editLink: string
+  editLink?: string
   item: T
-  onRemove: (t: T) => void
+  onRemove?: (t: T) => void
+}
+
+export interface ICRUDItemInfoProps<T> {
+  item: T
+  nameKey: keyof T
+}
+
+export class CRUDItemInfo<T>
+extends React.PureComponent<ICRUDItemInfoProps<T>> {
+  render() {
+    const {item, nameKey} = this.props
+    return <span>{item[nameKey]}</span>
+  }
 }
 
 export class CRUDItemRow<T> extends React.PureComponent<ICRUDItemRowProps<T>> {
   handleRemove = () => {
     const {onRemove, item} = this.props
-    onRemove(item)
+    if (onRemove) {
+      onRemove(item)
+    }
   }
   render() {
     const {nameKey, editLink, item} = this.props
@@ -31,23 +48,30 @@ export class CRUDItemRow<T> extends React.PureComponent<ICRUDItemRowProps<T>> {
     return (
       <React.Fragment>
         <div className='item-info'>
-          {item[nameKey]}
+          {this.props.Info
+            ? <this.props.Info item={item} nameKey={nameKey} />
+            : <CRUDItemInfo<T> item={item} nameKey={nameKey} />
+          }
         </div>
         <div className='ml-auto'>
-          <Link to={editLink}>
-            <Button isInverted isColor='link' aria-label='Edit'>
-              <FaEdit />
-            </Button>
-          </Link>
+          {!!editLink && (
+            <Link to={editLink}>
+              <Button isInverted isColor='link' aria-label='Edit'>
+                <FaEdit />
+              </Button>
+            </Link>
+          )}
           &nbsp;
-          <Button
-            aria-label='Remove'
-            onClick={this.handleRemove}
-            isColor='danger'
-            isInverted
-          >
-            <FaTimes />
-          </Button>
+          {!!this.props.onRemove && (
+            <Button
+              aria-label='Remove'
+              onClick={this.handleRemove}
+              isColor='danger'
+              isInverted
+            >
+              <FaTimes />
+            </Button>
+          )}
         </div>
       </React.Fragment>
     )
@@ -63,12 +87,14 @@ export class CRUDList<T> extends React.PureComponent<ICRUDListProps<T>> {
         <PanelHeading>
           <span className='is-flex v-centered'>
             <span>{title}</span>
-            <Link
-              className='ml-auto button is-link is-small'
-              to={newLink}
-            >
-              <FaPlus />&nbsp; New
-            </Link>
+            {!!newLink && (
+              <Link
+                className='ml-auto button is-link is-small'
+                to={newLink}
+              >
+                <FaPlus />&nbsp; New
+              </Link>
+            )}
           </span>
         </PanelHeading>
         {itemIds.map(itemId => {
@@ -76,8 +102,9 @@ export class CRUDList<T> extends React.PureComponent<ICRUDListProps<T>> {
           return (
             <PanelBlock key={itemId}>
               <CRUDItemRow<T>
+                Info={this.props.Info}
                 nameKey={nameKey}
-                editLink={editLink(item)}
+                editLink={editLink && editLink(item)}
                 item={item}
                 onRemove={this.props.onRemove}
               />
